Add Typst integration and comprehensive documentation
- Add WebUITypst module for Typst article generation with code blocks
- Add comprehensive DocC documentation with tutorials
- Improve CSS generator with enhanced class collection
- Add extensive test coverage for Typst and CSS features
- Remove legacy documentation files (README, CONTRIBUTING, etc.)
- Add GitHub Actions release workflow
- Fix workflow to use Ubuntu 22.04 and setup-swift@v2

+++ Sources/WebUI/Styles/Responsive/ResponsiveNamespace.swift
@@ -293,7 +293,7 @@ public enum ResponsiveStyle {
     public static func position(
         _ type: PositionType? = nil,
         at edges: Edge...,
-        offset length: Int? = nil
+        offset length: Double? = nil
     ) -> ResponsiveModification {
         let params = PositionStyleOperation.Parameters(
             type: type,
